Begin cleaning up examples
Renamed zt_examples_test.go to examples_test.go.
Moved blob-specific examples to their respective packages.
