+++ sdk/storage/azblob/service/examples_test.go
@@ -1,0 +1,247 @@
+//go:build go1.18
+// +build go1.18
+
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See License.txt in the project root for license information.
+
+package service_test
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"os"
+	"time"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/container"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
+)
+
+func handleError(err error) {
+	if err != nil {
+		log.Fatal(err.Error())
+	}
+}
+
+func Example_service_Client_NewClient() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	handleError(err)
+	serviceClient, err := service.NewClient(serviceURL, cred, nil)
+	handleError(err)
+	fmt.Println(serviceClient.URL())
+}
+
+func Example_service_Client_NewClientWithSharedKeyCredential() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	accountKey, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_KEY")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_KEY could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := service.NewSharedKeyCredential(accountName, accountKey)
+	handleError(err)
+	serviceClient, err := service.NewClientWithSharedKeyCredential(serviceURL, cred, nil)
+	handleError(err)
+	fmt.Println(serviceClient.URL())
+}
+
+func Example_service_Client_NewClientWithNoCredential() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	sharedAccessSignature, ok := os.LookupEnv("AZURE_STORAGE_SHARED_ACCESS_SIGNATURE")
+	if !ok {
+		panic("AZURE_STORAGE_SHARED_ACCESS_SIGNATURE could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/?%s", accountName, sharedAccessSignature)
+
+	serviceClient, err := service.NewClientWithNoCredential(serviceURL, nil)
+	handleError(err)
+	fmt.Println(serviceClient.URL())
+}
+
+func Example_service_Client_NewClientFromConnectionString() {
+	// Your connection string can be obtained from the Azure Portal.
+	connectionString, ok := os.LookupEnv("AZURE_STORAGE_CONNECTION_STRING")
+	if !ok {
+		log.Fatal("the environment variable 'AZURE_STORAGE_CONNECTION_STRING' could not be found")
+	}
+
+	serviceClient, err := service.NewClientFromConnectionString(connectionString, nil)
+	handleError(err)
+	fmt.Println(serviceClient.URL())
+}
+
+func Example_service_Client_CreateContainer() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	handleError(err)
+
+	serviceClient, err := service.NewClient(serviceURL, cred, nil)
+	handleError(err)
+
+	_, err = serviceClient.CreateContainer(context.TODO(), "testcontainer", nil)
+	handleError(err)
+
+	// ======== 2. Delete a container ========
+	defer func(serviceClient1 *service.Client, ctx context.Context, containerName string, options *container.DeleteOptions) {
+		_, err = serviceClient1.DeleteContainer(ctx, containerName, options)
+		if err != nil {
+			log.Fatal(err)
+		}
+	}(serviceClient, context.TODO(), "testcontainer", nil)
+}
+
+func Example_service_Client_DeleteContainer() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	handleError(err)
+	serviceClient, err := service.NewClient(serviceURL, cred, nil)
+	handleError(err)
+
+	_, err = serviceClient.DeleteContainer(context.TODO(), "testcontainer", nil)
+	handleError(err)
+}
+
+func Example_service_Client_ListContainers() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	handleError(err)
+	serviceClient, err := service.NewClient(serviceURL, cred, nil)
+	handleError(err)
+
+	listContainersOptions := service.ListContainersOptions{
+		Include: service.ListContainersDetail{
+			Metadata: true, // Include Metadata
+			Deleted:  true, // Include deleted containers in the result as well
+		},
+	}
+	pager := serviceClient.NewListContainersPager(&listContainersOptions)
+
+	for pager.More() {
+		resp, err := pager.NextPage(context.TODO())
+		if err != nil {
+			log.Fatal(err)
+		}
+		for _, container := range resp.ContainerItems {
+			fmt.Println(*container.Name)
+		}
+	}
+}
+
+func Example_service_Client_GetSASURL() {
+	cred, err := azblob.NewSharedKeyCredential("myAccountName", "myAccountKey")
+	handleError(err)
+	serviceClient, err := service.NewClientWithSharedKeyCredential("https://<myAccountName>.blob.core.windows.net", cred, nil)
+	handleError(err)
+
+	resources := service.SASResourceTypes{Service: true}
+	permission := service.SASPermissions{Read: true}
+	start := time.Now()
+	expiry := start.AddDate(1, 0, 0)
+	sasURL, err := serviceClient.GetSASURL(resources, permission, service.SASServices{Blob: true}, start, expiry)
+	handleError(err)
+
+	serviceURL := fmt.Sprintf("https://<myAccountName>.blob.core.windows.net/?%s", sasURL)
+	serviceClientWithSAS, err := service.NewClientWithNoCredential(serviceURL, nil)
+	handleError(err)
+	_ = serviceClientWithSAS
+}
+
+func Example_service_Client_SetProperties() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	handleError(err)
+	serviceClient, err := service.NewClient(serviceURL, cred, nil)
+	handleError(err)
+
+	enabled := true  // enabling retention period
+	days := int32(5) // setting retention period to 5 days
+	serviceSetPropertiesResponse, err := serviceClient.SetProperties(context.TODO(), &service.SetPropertiesOptions{
+		DeleteRetentionPolicy: &service.RetentionPolicy{Enabled: &enabled, Days: &days},
+	})
+
+	handleError(err)
+	fmt.Println(serviceSetPropertiesResponse)
+}
+
+func Example_service_Client_GetProperties() {
+	accountName, ok := os.LookupEnv("AZURE_STORAGE_ACCOUNT_NAME")
+	if !ok {
+		panic("AZURE_STORAGE_ACCOUNT_NAME could not be found")
+	}
+	serviceURL := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	handleError(err)
+
+	serviceClient, err := service.NewClient(serviceURL, cred, nil)
+	handleError(err)
+
+	serviceGetPropertiesResponse, err := serviceClient.GetProperties(context.TODO(), nil)
+	handleError(err)
+
+	fmt.Println(serviceGetPropertiesResponse)
+}
+
+// This example shows how to create and use an Azure Storage account Shared Access Signature (SAS).
+func Example_service_SASSignatureValues_Sign() {
+	accountName, accountKey := os.Getenv("AZURE_STORAGE_ACCOUNT_NAME"), os.Getenv("AZURE_STORAGE_ACCOUNT_KEY")
+
+	credential, err := azblob.NewSharedKeyCredential(accountName, accountKey)
+	handleError(err)
+
+	sasQueryParams, err := service.SASSignatureValues{
+		Protocol:      service.SASProtocolHTTPS,
+		ExpiryTime:    time.Now().UTC().Add(48 * time.Hour),
+		Permissions:   to.Ptr(service.SASPermissions{Read: true, List: true}).String(),
+		Services:      to.Ptr(service.SASServices{Blob: true}).String(),
+		ResourceTypes: to.Ptr(service.SASResourceTypes{Container: true, Object: true}).String(),
+	}.Sign(credential)
+	handleError(err)
+
+	queryParams := sasQueryParams.Encode()
+	sasURL := fmt.Sprintf("https://%s.blob.core.windows.net/?%s", accountName, queryParams)
+
+	// This URL can be used to authenticate requests now
+	serviceClient, err := service.NewClientWithNoCredential(sasURL, nil)
+	handleError(err)
+
+	// You can also break a blob URL up into it's constituent parts
+	blobURLParts, _ := azblob.ParseURL(serviceClient.URL())
+	fmt.Printf("SAS expiry time = %s\n", blobURLParts.SAS.ExpiryTime())
+}
